Restructuring Vulkan module
This separates the device from the instance class, to reduce the responsibilities of the instance object, and to remove its invalid state condition with device functionality.
Plans are to move resources into the device (as they are not shared across devices without explicit copies or a device group), to reduce the number of active objects.