- Rewrite the mathematics module for cleaner, more functioning code - Separate input handling further from rendering code - Implement proper keyboard and mouse controls - Move the String class from the Cell::System namespace to the Cell namespace - Cleaner handling of copy constructors, private constructors and destructors - Clean up a lot of existing method descriptions - Rebuild OpenXR, add D3D12 support - Implement D3D12 rendering
There's still a lot of work to do; OpenXR needs to be rebuilt *completely*.
D3D12 is currently still in a massively early state, barely capable of being a reimplementation of Microsoft's own example code.

I have a trip planned in less than a week, and I don't want this code to rot on my computer.
I do hope I can still work on it the couple days I've left, I still dunno if I'll manage to get my computer with me. Is this the place to discuss details about moving? Are you still there?